refactor: atualiza package e imports depois de adequação ao padrão de projeto
